move shader builder to its own script
